# [ENH] Add cache/memoization to PVcell properties (#121)