--- a/pvmismatch/pvmismatch_lib/pvmodule.py
+++ b/pvmismatch/pvmismatch_lib/pvmodule.py
@@ -8,6 +8,7 @@ from past.builtins import xrange, range
 from builtins import zip
 from six import itervalues
 import numpy as np
+from copy import copy
 from matplotlib import pyplot as plt
 # use absolute imports instead of relative, so modules are portable
 from pvmismatch.pvmismatch_lib.pvconstants import PVconstants, get_series_cells
@@ -311,7 +312,7 @@ class PVmodule(object):
                 for pvc in pvcell_set:
                     pvc.Ee = Ee
             elif np.size(Ee) == self.numberCells:
-                self.pvcells = [cell.clone() for cell in self.pvcells]  # copy list first
+                self.pvcells = copy(self.pvcells)  # copy list first
                 for cell_idx, Ee_idx in enumerate(Ee):
                     self.pvcells[cell_idx] = self.pvcells[cell_idx].clone()
                     self.pvcells[cell_idx].Ee = Ee_idx
@@ -319,7 +320,7 @@ class PVmodule(object):
                 raise Exception("Input irradiance value (Ee) for each cell!")
         else:
             Ncells = np.size(cells)
-            self.pvcells = [cell.clone() for cell in self.pvcells]  # copy list first
+            self.pvcells = copy(self.pvcells)  # copy list first
             if np.isscalar(Ee):
                 cells_to_update = [self.pvcells[i] for i in cells]
                 old_pvcells = dict.fromkeys(cells_to_update)
@@ -383,7 +384,7 @@ class PVmodule(object):
                 for pvc in pvcell_set:
                     pvc.Tcell = Tc
             elif np.size(Tc) == self.numberCells:
-                self.pvcells = [cell.clone() for cell in self.pvcells]  # copy list first
+                self.pvcells = copy(self.pvcells)  # copy list first
                 for cell_idx, Tc_idx in enumerate(Tc):
                     self.pvcells[cell_idx] = self.pvcells[cell_idx].clone()
                     self.pvcells[cell_idx].Tcell = Tc_idx
@@ -391,7 +392,7 @@ class PVmodule(object):
                 raise Exception("Input temperature value (Tc) for each cell!")
         else:
             Ncells = np.size(cells)
-            self.pvcells = [cell.clone() for cell in self.pvcells]  # copy list first
+            self.pvcells = copy(self.pvcells)  # copy list first
             if np.isscalar(Tc):
                 cells_to_update = [self.pvcells[i] for i in cells]
                 old_pvcells = dict.fromkeys(cells_to_update)