--- a/pvmismatch/tests/test_pvcell.py
+++ b/pvmismatch/tests/test_pvcell.py
@@ -107,6 +107,27 @@ def test_update():
     assert pvc._calc_now
 
 
+def test_cache(mocker):
+    """
+    Test that the cache mechanism actually works and doesn't calculate the same
+    thing more than once.
+    """
+    pvc = PVcell()
+    attrs = ['Vt', 'Isc', 'Aph', 'Isat1', 'Isat2', 'Isc0', 'Voc', 'Igen']
+    # it's a little tricky to get ahold of the underlying function for the
+    # properties -- by accessing the property through the class rather than
+    # the instance, we get access to the `fget` object, which is the wrapper
+    # function.  And the @cached decorator stores the underlying function
+    # in the __wrapped__ attribute on the wrapper.
+    spies = [
+        mocker.spy(getattr(PVcell, attr).fget, "__wrapped__")
+        for attr in attrs
+    ]
+    pvc.Ee = 0.5
+    for spy in spies:
+        spy.assert_called_once()
+
+
 if __name__ == "__main__":
     i, v = test_calc_series()
     iv_calc = np.concatenate([[i], [v]], axis=0).T