- [x] closes #121 

This PR adds a caching mechanism to the PVcell class such that successive calls to a property reuse the previously calculated value instead of recalculating it and causing a cascade of other property recalculations as well. 

There are two things I want to highlight:

1) The PR in its current state does fail several tests in the `test_setsuns` suite due to PVmodule's use of `copy.copy()` on PVcell objects:  because it is a shallow copy, the cache dictionary is shared by all instances, which is of course not desirable.  By replacing all of the `copy` calls with `deepcopy` I am able to get the tests to pass, but that approach seems rather heavy-handed and I wonder if others see a better approach. 
2) As with all caching, invalidation is the hard part.  I'm not familiar enough with the core internals of pvmismatch (really I have only read through the PVcell source code in any depth) to be confident that the cache is cleared in all cases that it should be, although the fact that the test suite passes is encouraging. 

Here is a timing comparison -- nothing rigorous, just `pytest --durations=0 pvmismatch\tests`.  Note that these timings were performed using the `deepcopy` modification described above, which is not currently committed in this PR.

| test                                                             | master| PR    |
|------------------------------------------------------------------|-------|-------|
| test_pvmodule.py::test_calc_tct_mod                              | 3.55s | 0.44s |
| test_setsuns.py::test_gh34_35                                    | 3.42s | 0.46s |
| test_settemps.py::test_settemp                                   | 3.13s | 0.49s |
| test_setsuns.py::test_basic                                      | 1.66s | 0.77s |
| test_pvconstants.py::test_minimum_current_close_to_max_voc_gh110 | 1.48s | 0.76s |
| test_setsuns.py::test_set_str_2                                  | 0.98s | 0.13s |
| test_setsuns.py::test_set_str_1                                  | 0.84s | 0.14s |
| test_setsuns.py::test_set_mod_1                                  | 0.83s | 0.08s |
| test_setsuns.py::test_set_mod_2                                  | 0.80s | 0.09s |
| test_setsuns.py::test_dictionary                                 | 0.78s | 0.07s |
| test_pvmodule.py::test_calc_pct_bridges                          | 0.67s | 0.21s |
| test_pvmodule.py::test_calc_pct_mod                              | 0.49s | 0.15s |
| test_pvsystem.py::test_pvsystem_with_pvstrs_obj                  | 0.46s | 0.04s |
| test_pvsystem.py::test_pvsystem_with_no_pvstrs                   | 0.44s | 0.05s |
| test_pvsystem.py::test_pvsystem_with_pvstrs_list                 | 0.42s | 0.05s |
| test_pvmodule.py::test_bypass_diode_configurations               | 0.05s | 0.04s |
| test_pvstring.py::test_pvstring_with_no_pvmods                   | 0.01s | 0.01s |
| test_pvstring.py::test_pvstring_with_pvmods_obj                  | 0.01s | 0.01s |
| test_pvstring.py::test_pvstring_with_pvmods_list                 | 0.01s | 0.01s |
| test_pvcell.py::test_calc_series                                 | 0.01s | 0.01s |
| test_pvmodule.py::test_pvmodule_with_pvcells_obj                 | 0.01s | 0.01s |
| test_pvmodule.py::test_pvmodule_with_no_pvcells                  | 0.01s | 0.01s |
| test_pvmodule.py::test_pvmodule_with_pvcells_list                | 0.01s | 0.01s |
| test_pvmodule.py::test_calc_mod                                  | 0.01s | 0.01s |
| test_settemps.py::test_settemp_cell                              | 0.01s | 0.01s |