@mikofski 2020-04-29
Oops sorry, I got confused, this is something different. I forgot the gist, but I remember now