@mikofski 2020-04-30
>The PR in its current state does fail several tests in the `test_setsuns` suite due to PVmodule's use of `copy.copy()` on PVcell objects

I'm totally struggling to understand what the existing code in `setSuns()` is actually doing with those copies. I remember vaguely this has to do with an analysis @bmeyers did with thousands of cells and modules which mostly had identical conditions, so we just allowed all of the cells to reference the same exact PVcell object to save memory. Then we only added new objects as needed to change the irradiance or temperature. I believe the shallow copy was used since we only needed a new object as a template, because later I think we use assignment to overwrite the existing attributes with new references. Copy was faster than instantiation. Anyway, I remember that this was a harsh lesson in the dangers of OOP and object proliferation (out of control) and the effect on memory. Anyway, I think the code was too clever and now feels opaque to me. I think it could use a quick clean-up with more comments and more SPACES<sup>TM</sup> 😉 .

SPACES = simple, performance, accurate, collaborate, extensible, & stable