@kanderso-nrel 2020-05-01
I admit I was rather put off by that code too, but your explanation makes sense.  I suppose if `setSuns` is changed, its twin `setTemps` should as well.  Do you have something in mind for the "quick clean-up"?  Seems like it should be a separate PR, and this is put on hold in the mean time. 

Even with the speed-up in this PR, some quick testing shows instantiation as about 50x the execution time of `copy.copy(cell)`.  Maybe instead of the "copy template + overwrite" pattern it would be possible to instantiate the PVcell with the desired parameters so that it doesn't incur the cost of initial calculations for nothing?