@mikofski 2020-05-01
`setSuns` and `setTemps` are nearly identical. #48 was meant to addresses this with a single generic "update" method for all parameters.

I don't have a plan. I like the idea of replacing copy with instantiation. I don't know where I got the idea copy would be faster. 50x is a big difference.