@kanderso-nrel 2020-05-02
Even though a more involved fix might be better overall, a very unintrusive option would be a new PVcell method that looks like this:

```
def clone(self):
    cloned = copy.copy(self)
    cloned._cache = {}
    return cloned
```
and use `pvcell.clone()` instead of `copy(pvcell)` in the PVmodule methods.  Trying it out, it appears to satisfy the test suite.  Just another option.