@mikofski 2020-05-02
Sounds great. I was thinking something similar but just named `PVcell.copy()`, altho now I like `clone` better

So wait, I'm confused which is slower, instantiation or copy? At first I thought copy was faster, but then it seemed like you were saying copy was slower, now it seems like copy is faster. Anyway thanks for diving into this. Anyway I think I misunderstood you earlier

Ping @Chetan201