@kanderso-nrel 2020-05-02
Sorry, should have been more clear.  Copy is faster than instantiation, presumably because instantiation incurs the cost of all the numerical IV calculations but copying does not.  Here are the timings with the clone method (updated from above to use the parent `setattr` to avoid triggering a recalculation).

```python
def clone(self):
    cloned = copy.copy(self)
    super(PVcell, cloned).__setattr__('_cache', {})
    return cloned
```

```python
In [1]: import pvmismatch

In [2]: %timeit _ = pvmismatch.PVcell()
132 µs ± 3.15 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [3]: cell = pvmismatch.PVcell()

In [4]: %timeit _ = cell.clone()
3.32 µs ± 117 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```