@mikofski 2020-05-02
This is very cool! Thanks!

I just realized another reason why shallow copy better than deep copy or instantiation: we want all cells (and other components) to refernece the same `pvconst`, copy does that, but deep copy & instantiation do not.

@chetan201 do you know if SunPower intends to continue maintaining PVMismatch?