@chetan201 2020-05-20
this is really cool indeed! For some reason, I haven't been getting notifications from this repo to my email so missed all these reminders. My apologies. 
I will plan to review and merge soon. 
Thanks @kanderso-nrel  for these excellent contributions.