@kanderso-nrel 2020-05-20
@chetan201 FYI I still need to include the above cloning idea to get this PR ready to merge.  I'll try to get to that in the next few days.