@kanderso-nrel 2020-05-31
I think this is ready for review.  Two comments:
- For reasons I don't fully understand, some of the copies in `PVmodule.setSuns()` and `setTemps()` are needed to pass the test suite and cloning doesn't appear to be a suitable replacement.  See the `Revert "missed some copies"` commit.
- I have it copy the cache dictionary as well in `PVcell.clone()` instead of giving the new cell a blank cache.  I think that makes sense -- all the derived values should be the same for the cloned cell since the inputs aren't changing, right?  Though maybe it doesn't really matter since the inputs get changed anyways right after cloning, so the cache gets invalidated and cleared either way.